python-backend/requirements.txt
python-dotenv
livekit-plugins-google
opencv-python
